+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - `setup_tbs` as the public interface
     - underscore keys in the subscription definitions
     - non-managed subscriptions by default
+    - new way of subscription definition
 - Support for non-managed subscriptions
 - Support for optional token auth for webhook
 - Support for both `id-only`/`full-resource`